requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
backports.functools-lru-cache==1.6.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
configparser==4.0.2
contextlib2==0.6.0.post1
crcmod==1.7
cryptography==3.3.2
enum34==1.1.10
future==0.18.2
funcsigs==1.0.2
idna==2.10
importlib-metadata==2.1.1
ipaddress==1.0.23
jmespath==0.10.0
more-itertools==5.0.0
packaging==20.9
pathlib2==2.3.5
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
pytest==4.6.11
requests==2.25.1
requests-toolbelt==0.9.1
scandir==1.10.0
simplejson==3.17.2
six==1.15.0
urllib3==1.26.4
wcwidth==0.2.5
zipp==1.2.0
